ref: separate generator code from otp
